Remove K_CTYPES; represent structs/arrays in Type (#144)
* WIP: Remove K_CTYPES; represent structs/arrays in Type

* Update docs, use array_dim, normalize get_deref_paths

* Small cleanup before review

* Remove hack from late_field_path

* Incorporate PR feedback from Simon:

- Add clarifying inline comments
- Do not treat arrays as pointers in `.get_pointer_target()`/`.get_deref_field()`
- Support arrays with unknown dimensions
- Change `Type.get_field()` return type
- Change `TypeUniverse.add_struct_type()` signature
- Rename `.has_late_field_path` -> `.checked_late_field_path`
- Rename `.is_function_void()` -> `.is_function_known_void()`

* Rename Type.size -> Type.size_bits

* Refactor StructDeclaration to use byte sizes/offsets/alignments

* Narrow `TypeMap.struct_typedefs` type; remove prev TODO

* Check `sub_remainder_bits == 0`

This changes 2 OOT tests (which I think is an improvement):

```diff
$ ./run_tests.py --project-with-context ../oot  -j4 --diff-context=0 2>| grep -v PASS
[INFO] Overwriting test output files.
[INFO] [FAIL] oot_ctx:non_matchings/code/audio_seqplayer/Audio_SequenceChannelProcessScript.s
[INFO] Output of ../oot/asm/non_matchings/code/audio_seqplayer/Audio_SequenceChannelProcessScript.s changed! Diff:
---

+++

@@ -270 +270 @@

-                        arg0->unk2 = sp63;
+                        arg0->noteAllocPolicy = sp63;
@@ -294 +294 @@

-                        arg0->unk2 = (u8) temp_s0_6->unk-2;
+                        arg0->noteAllocPolicy = temp_s0_6->unk-2;
@@ -306 +306 @@

-                        arg0->unk2 = (u8) sp64;
+                        arg0->noteAllocPolicy = (u8) sp64;
[INFO] [FAIL] oot_ctx:non_matchings/code/code_800DACC0/func_800DC384.s
[INFO] Output of ../oot/asm/non_matchings/code/code_800DACC0/func_800DC384.s changed! Diff:
---

+++

@@ -120 +120 @@

-                        if ((((u32) phi_v1->unk0 >> 0x1F) != 0) && (phi_s4 == ((u32) phi_v1->unk1 >> 5))) {
+                        if ((((u32) phi_v1->unk0 >> 0x1F) != 0) && (phi_s4 == ((u32) phi_v1->bitField1.asByte >> 5))) {
@@ -146 +146 @@

-                if ((((u32) phi_v1_2->unk0 >> 0x1F) != 0) && ((s32) ((u32) phi_v1_2->unk1 >> 5) >= (s32) temp_t1)) {
+                if ((((u32) phi_v1_2->unk0 >> 0x1F) != 0) && ((s32) ((u32) phi_v1_2->bitField1.asByte >> 5) >= (s32) temp_t1)) {
@@ -218 +218 @@

-                if (phi_s4_2 == ((u32) temp_a1->unk1 >> 5)) {
+                if (phi_s4_2 == ((u32) temp_a1->bitField1.asByte >> 5)) {
[INFO] Test summary: 1719 passed, 0 skipped, 2 failed, 1721 total
```

* Unify types in `StructAccess.late_field_path()`

Test output diff from previous commit here:

https://gist.github.com/zbanks/4e0bee48a841799a01b65f2496fa26ed

* PR feedback from Simon

- Remove references to "bits" in `Type.get_field()`
- Simplify `Type.get_deref_field()`
- Remove assertion for `array_dim is not None`
- Change `Type.ctype()` behavior to set `array_dim=None` for decls like `x[];`

Diff in OOT corpus:
```diff
$ ./run_tests.py --project-with-context ../oot -j4 --diff-context=1 2>| grep -v PASS
[INFO] Output of ../oot/asm/non_matchings/code/audio_seqplayer/Audio_SequenceChannelProcessScript.s changed! Diff:
---
+++
@@ -359,3 +359,3 @@
                     case 182: // switch 1
-                        temp_s6->unk19 = (s8) *(arg0->dynTable + temp_s6->unk19);
+                        temp_s6->unk19 = (s8) arg0->dynTable[0][0][temp_s6->unk19];
                         goto loop_7;
[INFO] Test summary: 1720 passed, 0 skipped, 1 failed, 1721 total
```

* Rework `StructDeclaration.from_ctype()` API

* Respond to Simon's PR comments for translate.py

- Clarify/add docs
- Add `static_assert_unreachable` to error.py
- `del x[-1]` -> `x.pop()`
- `Type.get_size_align_bytes()` -> `Type.get_parameter_size_align_bytes()`
- Clean up `StructAccess.format()`; make rewrite rules explicit

* Merge GlobalInfo & TypeUniverse

- Remove TypeUniverse references from Type
- Delete .get_var_type() and .get_function_type
- Create GlobalTypeInfo interface to avoid circular refs
- One test changed where a function was only used as a function pointer

* Respond to PR comments from Simon

Rework `Type.get_field()` usage & `StructAccess.field_path`s. Now,
`Type.get_field()` is not used at all from `translate.py`, and all
`StructAccess.field_path` values start with an `int`.

* Check `len(field_path) >= 2` in `StructAccess.make_reference()`

* Separate out TypePool from GlobalInfo

* Refactor ctype maps to avoid `id(...)`

* Small PR feedback from Simon

- Add `StructAccess.__post_init__()`
- Add `Type.decay()`
- Add comments

* Refactor StructAccess field_path invariant checks

+++ tests/end_to_end/if_postdec/irix-g-out.c
@@ -4,7 +4,7 @@ extern s32 D_4100F0;
     s32 temp_t6;
 
     temp_t6 = D_4100F0;
-    D_4100F0 = (s32) (temp_t6 - 1);
+    D_4100F0 = temp_t6 - 1;
     if (temp_t6 < 1) {
         return 4;
     }
