+++ tests/end_to_end/lwl/irix-o2-out.c
@@ -17,5 +17,5 @@ void test(void) {
     D_410161 = (unaligned s32) D_410169;
     D_410170.unk0 = (unaligned s32) D_410160.unk0;
     D_410170.unk4 = (u8) D_410160.unk4;
-    D_410178 = (s32) (unaligned s32) D_400158;
+    D_410178 = (unaligned s32) D_400158;
 }
